modularised the cli.py
